% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ci.R
\name{calc_ci}
\alias{calc_ci}
\title{Calculate performance criteria and MCSE.}
\usage{
calc_ci(
  lower_bound = NULL,
  upper_bound = NULL,
  true_param,
  alpha = 0.05,
  K,
  perfm_criteria = c("coverage", "width")
)
}
\arguments{
\item{lower_bound}{A numeric vector indicating the lower bound for confidence interval coverage and width calculations.}

\item{upper_bound}{A numeric vector indicating the upper bound for confidence interval coverage and width calculations.}

\item{true_param}{A number indicating the true parameter.}

\item{alpha}{A number indicating the nominal alpha level.}

\item{K}{A number indicating number of simulation iterations.
@param perfm_criteria A character or a character vector indicating the performance criteria to be calculated.}
}
\value{
The performance criteria estimate and the associated MCSE.
}
\description{
Calculate performance criteria and MCSE.
}
