% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_coverage.R
\name{calc_coverage}
\alias{calc_coverage}
\title{Calculate performance criteria and MCSE.}
\usage{
calc_coverage(
  cov_dat,
  true_param,
  alpha = 0.05,
  K,
  perfm_criteria = c("coverage", "width")
)
}
\arguments{
\item{cov_dat}{A dataframe or tibble containing two columns called lower_bound and upper_bound containing the confindence interval lower and upper bounds.}

\item{true_param}{A number indicating the true parameter.}

\item{alpha}{A number indicating the nominal alpha level.}

\item{K}{A number indicating number of simulation iterations.
@param perfm_criteria A character or a character vector indicating the performance criteria to be calculated.}
}
\value{
The performance criteria estimate and the associated MCSE.
}
\description{
Calculate performance criteria and MCSE.
}
\examples{
calc_coverage(cov_dat = t_res, true_param = .5, K = nrow(t_res))


}
