% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_jacknife.R
\name{calc_jacknife}
\alias{calc_jacknife}
\title{Calculate jacknife Monte Carlo SE for variance estimators.}
\usage{
calc_jacknife(res_dat, estimates, true_param, K)
}
\arguments{
\item{res_dat}{A dataframe or tibble containing variance estimates.}

\item{estimates}{The column containing the variance estimates.}

\item{true_param}{A number indicating the true parameter.}

\item{K}{A number indicating number of simulation iterations.}
}
\value{
The performance criteria estimate and the associated MCSE.
}
\description{
Calculate jacknife Monte Carlo SE for variance estimators.
}
\examples{
calc_jacknife(res_dat = alpha_res, estimates = var_est, true_param = .8, K = nrow(alpha_res))

}
