% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rbvjk.R
\name{calc_rbvjk}
\alias{calc_rbvjk}
\title{Calculate jacknife Monte Carlo SE for variance estimators.}
\usage{
calc_rbvjk(res_dat, estimates, var_estimates, true_param)
}
\arguments{
\item{res_dat}{data frame or tibble containing the simulation results.}

\item{estimates}{name of the column containing the estimates.}

\item{var_estimates}{name of the column containing the variance estimates.}

\item{true_param}{name of the column containing true parameters of the point estimator.}
}
\value{
A tibble containing the performance criteria estimate and the associated MCSE.
}
\description{
#' @description Calculates relative bias of variance estimators.
The function also calculates the associated jacknife Monte Carlo Standard errors.
}
\examples{
calc_rbvjk(res_dat = alpha_res, estimates = A, var_estimates = Var_A, true_param = true_param)

}
