% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_relative.R
\name{calc_relative}
\alias{calc_relative}
\title{Calculate performance criteria and MCSE.}
\usage{
calc_relative(
  estimates,
  true_param,
  K,
  perfm_criteria = c("relative bias", "relative mse")
)
}
\arguments{
\item{estimates}{A numeric vector including the estimates.}

\item{true_param}{A number indicating the true parameter.}

\item{K}{A number indicating number of simulation iterations.}

\item{perfm_criteria}{A character or a character vector indicating the performance criteria to be calculated.}
}
\value{
The performance criteria estimate and the associated MCSE.
}
\description{
Calculate performance criteria and MCSE.
}
