% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rr.R
\name{calc_rr}
\alias{calc_rr}
\title{Calculate performance criteria and MCSE.}
\usage{
calc_rr(rr_dat, alpha = 0.05, K)
}
\arguments{
\item{rr_dat}{A dataframe or tibble containing a column called p_vals containing p values}

\item{alpha}{A number indicating the nominal alpha level.}

\item{K}{A number indicating number of simulation iterations.}
}
\value{
The performance criteria estimate and the associated MCSE.
}
\description{
Calculate performance criteria and MCSE.
}
\examples{
calc_rr(rr_dat = t_res, K = nrow(t_res))


}
