% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_skeleton.R
\name{create_skeleton}
\alias{create_skeleton}
\title{Open a simulation skeleton}
\usage{
create_skeleton(filename, ...)
}
\arguments{
\item{filename}{Name of file to create.}

\item{...}{Further arguments passed to file.copy}
}
\description{
Creates and opens an .R file containing a skeleton for writing a Monte Carlo simulation study.
}
\examples{
\dontrun{
create_skeleton("Behrens_Fisher_problem")
}
}
