% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial_summary.R
\name{factorial_summary}
\alias{factorial_summary}
\title{Summarize the sums of squares attributable to each factor in a full
factorial experimental design}
\usage{
factorial_summary(
  data,
  y,
  factors,
  sum_orders = TRUE,
  include_total = TRUE,
  check_balance = TRUE
)
}
\arguments{
\item{data}{data.frame or tibble containing simulation results. Each row
should correspond to a unique set of parameter values.}

\item{y}{character string corresponding to the outcome variable in
\code{data}.}

\item{factors}{character vector containing the names of two or more variables
in \code{data} that correspond to factors in the experimental design.}

\item{sum_orders}{logical indicating whether to compute the total sum of squares
attributable to each factor, with a default of \code{TRUE}.}

\item{include_total}{logical indicating whether to compute the total sum of squares
for each term order (across all factors), with a default of \code{TRUE}.}

\item{check_balance}{logical indicating whether to check that the experimental design is balanced, with a default of \code{TRUE}.}
}
\value{
A data.frame
}
\description{
Computes an analysis of variance for a set of factors in a full
factorial experimental design. For each term order within the design (i.e.,
main effects, two-way interactions, three-way interactions), the total sum
of squares attributable to each factor is computed. For terms beyond order
\enumerate{
\item Optionally, also computes the total sum of squares for each order
(across all factors) and the total sum of squares attributable to each
factor.
}
}
\examples{

data("Chen_Pusto")
dat <- subset(Chen_Pusto, method == "PET-PEESE")

factorial_summary(dat, "bias", c("k","mu","tau","cor_mu","wts"))
factorial_summary(dat, "bias", c("k","mu","tau","wts"), include_total = FALSE)

}
