% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_coverage.R
\name{project_coverage}
\alias{project_coverage}
\title{Extrapolate coverage and width using sub-sampled bootstrap confidence
  intervals.}
\usage{
project_coverage(
  data,
  CI_subsamples,
  true_param,
  criteria = c("coverage", "width"),
  winz = Inf
)
}
\arguments{
\item{data}{data frame or tibble containing the simulation results.}

\item{CI_subsamples}{list or name of column from \code{data} containing list
of confidence intervals calculated based on sub-samples with different
numbers of replications.}

\item{true_param}{vector or name of column from \code{data} containing
corresponding true parameters.}

\item{criteria}{character or character vector indicating the performance
criteria to be calculated, with possible options \code{"coverage"} and
\code{"width"}.}

\item{winz}{numeric value for winsorization constant. If set to a finite
value, estimates will be winsorized at the constant multiple of the
inter-quartile range below the 25th percentile or above the 75th percentile
of the distribution. For instance, setting \code{winz = 3} will
truncate estimates that fall below P25 - 3 * IQR or above P75 + 3 * IQR.}
}
\value{
A tibble containing the number of simulation iterations, performance
  criteria estimate(s) and the associated MCSE.
}
\description{
Given a set of bootstrap confidence intervals calculated across
  sub-samples with different numbers of replications, extrapolates confidence
  interval coverage and width of bootstrap confidence intervals to a
  specified (larger) number of bootstraps. The function also calculates the
  associated Monte Carlo standard errors. The confidence interval percentage
  is based on how you calculated the lower and upper bounds.
}
\examples{
project_coverage(
  data = t_boots,
  CI_subsamples = CIs,
  true_param = true_param
)

}
