% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_abs.R
\name{calc_abs}
\alias{calc_abs}
\title{Calculate absolute performance criteria and MCSE.}
\usage{
calc_abs(
  res_dat,
  estimates,
  true_param,
  perfm_criteria = c("bias", "variance", "mse", "rmse")
)
}
\arguments{
\item{res_dat}{A dataframe or tibble containing simulation results.}

\item{estimates}{The name of the column containing estimates.}

\item{true_param}{The name of the column containing true parameters.}

\item{perfm_criteria}{A character or a character vector indicating the performance criteria to be calculated.}
}
\value{
A tibble containing the performance criteria estimate(s) and the associated MCSE.
}
\description{
Calculates absolute bias, variance, mean squared error (mse)
and root mean squared error (rmse). The function also calculates the associated
Monte Carlo Standard errors.
}
\examples{
calc_abs(res_dat = t_res, estimates = est, true_param = true_param)


}
