% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_abs.R
\name{calc_abs}
\alias{calc_abs}
\title{Calculate performance criteria and MCSE.}
\usage{
calc_abs(
  res_dat,
  estimates,
  true_param,
  perfm_criteria = c("bias", "variance", "mse", "rmse")
)
}
\arguments{
\item{res_dat}{A dataframe or tibble containing simulation results.}

\item{estimates}{The name of the column containing estimates.}

\item{true_param}{A number indicating the true parameter.}

\item{perfm_criteria}{A character or a character vector indicating the performance criteria to be calculated.}
}
\value{
The performance criteria estimate and the associated MCSE.
}
\description{
Calculate performance criteria and MCSE.
}
\examples{
calc_abs(res_dat = t_res, estimates = est, true_param = .5)


}
