% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_coverage.R
\name{calc_coverage}
\alias{calc_coverage}
\title{Calculate performance criteria and MCSE.}
\usage{
calc_coverage(
  res_dat,
  lower_bound,
  upper_bound,
  true_param,
  alpha = 0.05,
  perfm_criteria = c("coverage", "width")
)
}
\arguments{
\item{res_dat}{A dataframe or tibble containing confidence interval results.}

\item{lower_bound}{The column containing the lower bound estimates of the confidence intervals.}

\item{upper_bound}{The column containing the upper bound estimates of the confidence intervals.}

\item{true_param}{A number indicating the true parameter.}

\item{alpha}{A number indicating the nominal alpha level.}

\item{perfm_criteria}{A character or a character vector indicating the performance criteria to be calculated.}
}
\value{
The performance criteria estimate and the associated MCSE.
}
\description{
Calculate performance criteria and MCSE.
}
\examples{
calc_coverage(res_dat = t_res, lower_bound = lower_bound, upper_bound = upper_bound, true_param = .5)


}
