% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_var.R
\name{calc_jacknife}
\alias{calc_jacknife}
\title{Calculate jacknife Monte Carlo SE for variance estimators.}
\usage{
calc_jacknife(var_dat, true_param, K)
}
\arguments{
\item{true_param}{A number indicating the true parameter.}

\item{K}{A number indicating number of simulation iterations.}

\item{dat}{A dataframe or tibble containing a column called var_est - variance estimates.}
}
\value{
The performance criteria estimate and the associated MCSE.
}
\description{
Calculate jacknife Monte Carlo SE for variance estimators.
}
\examples{
\dontrun{
calc_jacknife(var_dat = var_res, true_param = .5, K = nrow(var_res))
}

}
