% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rbvjk.R
\name{calc_rbvjk}
\alias{calc_rbvjk}
\title{Calculate jacknife Monte Carlo SE for variance estimators.}
\usage{
calc_rbvjk(res_dat, estimates, var_estimates, true_param)
}
\arguments{
\item{res_dat}{data frame or tibble containing variance estimates.}

\item{estimates}{name of the column containing the estimates.}

\item{var_estimates}{name of the column containing the variance estimates.}

\item{true_param}{name of the column containing true parameters.}
}
\value{
A tibble containing the performance criteria estimate and the associated MCSE.
}
\description{
Calculate jacknife Monte Carlo SE for variance estimators.
}
\examples{
calc_rbvjk(res_dat = alpha_res, estimates = est, var_estimates = var_est, true_param = true_param)

}
