% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_relative.R
\name{calc_relative}
\alias{calc_relative}
\title{Calculate relative performance criteria and MCSE.}
\usage{
calc_relative(
  res_dat,
  estimates,
  true_param,
  perfm_criteria = c("relative bias", "relative mse")
)
}
\arguments{
\item{res_dat}{data frame or tibble containing the simulation results.}

\item{estimates}{name of the column containing the estimates}

\item{true_param}{name of the column containing the true parameters.}

\item{perfm_criteria}{character or character vector indicating the performance criteria to be calculated.}
}
\value{
A tibblecontaining the performance criteria estimate(s) and the associated MCSE.
}
\description{
Calculates relative bias, mean squared error (mse)
and root mean squared error (rmse). The function also calculates the associated
Monte Carlo Standard errors.
}
\examples{
calc_relative(res_dat = t_res, estimates = est, true_param = true_param)


}
