% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_relative.R
\name{calc_relative}
\alias{calc_relative}
\title{Calculate relative performance criteria and MCSE.}
\usage{
calc_relative(
  res_dat,
  estimates,
  true_param,
  perfm_criteria = c("relative bias", "relative mse")
)
}
\arguments{
\item{res_dat}{A dataframe or tibble containing a column called est - estimates}

\item{estimates}{The name of the column containing estimates}

\item{true_param}{The name of the column containing true parameters.}

\item{perfm_criteria}{A character or a character vector indicating the performance criteria to be calculated.}
}
\value{
A tibblecontaining the performance criteria estimate(s) and the associated MCSE.
}
\description{
Calculates relative bias, mean squared error (mse)
and root mean squared error (rmse). The function also calculates the associated
Monte Carlo Standard errors.
}
\examples{
calc_relative(res_dat = t_res, estimates = est, true_param = true_param)


}
