% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_relative.R
\name{calc_relative}
\alias{calc_relative}
\title{Calculate performance criteria and MCSE.}
\usage{
calc_relative(
  res_dat,
  estimates,
  true_param,
  perfm_criteria = c("relative bias", "relative mse")
)
}
\arguments{
\item{res_dat}{A dataframe or tibble containing a column called est - estimates}

\item{estimates}{The name of the column containing estimates}

\item{true_param}{A number indicating the true parameter.}

\item{perfm_criteria}{A character or a character vector indicating the performance criteria to be calculated.}
}
\value{
The performance criteria estimate and the associated MCSE.
}
\description{
Calculate performance criteria and MCSE.
}
\examples{
calc_relative(res_dat = t_res, estimates = est, true_param = .5)


}
