% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rr.R
\name{calc_rr}
\alias{calc_rr}
\title{Calculate rejection rate and MCSE.}
\usage{
calc_rr(res_dat, p_values, alpha = 0.05)
}
\arguments{
\item{res_dat}{A dataframe or tibble containing simulation results.}

\item{p_values}{The name of the column containing p values.}

\item{alpha}{A number indicating the nominal alpha level.}
}
\value{
A dataframe containing the  performance criteria estimate and the associated MCSE.
}
\description{
Calculates rejection rate. The function also calculates the associated
Monte Carlo Standard error.
}
\examples{
calc_rr(res_dat = t_res, p_values = p_val)


}
