% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rr.R
\name{calc_rr}
\alias{calc_rr}
\title{Calculate rejection rate and MCSE.}
\usage{
calc_rr(res_dat, p_values, alpha = 0.05)
}
\arguments{
\item{res_dat}{data frame or tibble containing the simulation results.}

\item{p_values}{name of the column containing the p values.}

\item{alpha}{number indicating the nominal alpha level. Default value is set to the conventional .05.}
}
\value{
A tibble containing the performance criteria estimate and the associated MCSE.
}
\description{
Calculates rejection rate. The function also calculates the associated
Monte Carlo Standard error.
}
\examples{
calc_rr(res_dat = t_res, p_values = p_val)


}
