% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rr.R
\name{calc_rr}
\alias{calc_rr}
\title{Calculate performance criteria and MCSE.}
\usage{
calc_rr(p_vals, alpha = 0.05, K)
}
\arguments{
\item{p_vals}{A numeric vector including the estimates.}

\item{alpha}{A number indicating the nominal alpha level.}

\item{K}{A number indicating number of simulation iterations.}
}
\value{
The performance criteria estimate and the associated MCSE.
}
\description{
Calculate performance criteria and MCSE.
}
