% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_res.R
\docType{data}
\name{t_res}
\alias{t_res}
\title{t-test Simulation Results}
\format{}
\usage{
t_res
}
\description{
A dataset containing simulation results from a study that just runs a t-test
}
\keyword{datasets}
